+++ fivox/compartmentLoader.cpp
@@ -53,9 +53,10 @@ public:
         helpers::addCompartmentEvents( morphologies, _report, output );
     }
 
-    ssize_t load( const float time )
+    ssize_t load()
     {
-        const brion::floatsPtr values = _report.loadFrame( time );
+        const brion::floatsPtr values =
+                _report.loadFrame( _output.getCurrentTime( ));
         if( !values )
             return -1;
 
@@ -86,9 +87,10 @@ Vector2f CompartmentLoader::_getTimeRange() const
                      _impl->_report.getEndTime( ));
 }
 
-ssize_t CompartmentLoader::_load( const float time )
+ssize_t CompartmentLoader::_load( const size_t /*chunkIndex*/,
+                                  const size_t /*numChunks*/ )
 {
-    return _impl->load( time );
+    return _impl->load();
 }
 
 }